backend/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
motor==3.1.1
pydantic==1.10.8
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
email-validator==1.3.1
python-dotenv==0.21.1
httpx==0.23.3
google-auth==2.16.2
openai==0.27.0
